1、strkit增加获取字符串长度函数。 2、IntIsZero函数参数类型由int64修改成int。

+++ intkit/IntKit.go
@@ -5,7 +5,7 @@ package intkit
 // @param strs
 // usage:
 //	IntIsZero(0) return false
-func IntIsZero(ints ...int64) bool {
+func IntIsZero(ints ...int) bool {
 	if len(ints) == 0 {
 		return false
 	}
